Sending additional valuable info in post run
